rename files according to criteria
