+++ src/App.tsx/App.tsx
@@ -1,13 +1,13 @@
-import FavouritesPage from '../pages/favorites-page/FavoritesPage';
-import LoginPage from '../pages/login-page/LoginPage';
-import MainPage from '../pages/main-page/MainPage';
+import FavouritesPage from '../pages/favorites-page/favorites-page';
+import LoginPage from '../pages/login-page/login-page';
+import MainPage from '../pages/main-page/main-page';
 import { placesToStay } from '../libs/mock-data';
 import {BrowserRouter, Routes, Route} from 'react-router-dom';
-import NotFoundPage from '../pages/not-found-page/NotFoundPage';
-import OfferPage from '../pages/offer-page/OfferPage';
+import NotFoundPage from '../pages/not-found-page/not-found-page';
+import OfferPage from '../pages/offer-page/offer-page';
 import Layout from '../layout/Layout';
 import { AppRoutes, AuthorisationStatus } from '../libs/const';
-import PrivateRoute from '../components/private-route/PrivateRoute';
+import PrivateRoute from '../components/private-route/private-route';
 
 function App() {
   return (
