+++ src/components/offer-card-main/offer-card.tsx
