+++ src/components/offer/near-places/near-places.tsx
