+++ src/pages/login-page/login-page.tsx
@@ -1,4 +1,4 @@
-import Logo from '../../components/logo/Logo';
+import Logo from '../../components/logo/logo';
 
 function LoginPage() {
   return (
